feat(quaint): enable pure Wasm SqliteError
