+++ quaint/src/connector/sqlite/wasm/ffi.rs
@@ -1,0 +1,7 @@
+//! This is a partial copy of `rusqlite::ffi::*`.
+pub const SQLITE_BUSY: i32 = 5;
+pub const SQLITE_CONSTRAINT_FOREIGNKEY: i32 = 787;
+pub const SQLITE_CONSTRAINT_NOTNULL: i32 = 1299;
+pub const SQLITE_CONSTRAINT_PRIMARYKEY: i32 = 1555;
+pub const SQLITE_CONSTRAINT_TRIGGER: i32 = 1811;
+pub const SQLITE_CONSTRAINT_UNIQUE: i32 = 2067;
