+++ quaint/src/connector/sqlite/wasm/mod.rs
@@ -2,3 +2,4 @@
 //! This module is only available with the `sqlite` feature.
 pub(crate) mod common;
 pub mod error;
+mod ffi;
